model relation

+++ 05_Django/05_django_form/articles/views.py
@@ -23,7 +23,10 @@ def create(request):
             # title = form.cleaned_data.get('title')
             # content = form.cleaned_data.get('content')
             # article = Article.objects.create(title=title, content=content)
-            article = form.save()
+            # article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
         return redirect('articles:detail', article.pk)
     else:
         form = ArticleForm()
@@ -31,7 +34,7 @@ def create(request):
     # 1. GET요청 -> 비어있는 폼 전달
     # 2. 유효성 검증 실패 -> 에러 메시지를 포함한 채로 폼 전달
     context = {'form':form}
-    return render(request, 'articles/create.html', context)
+    return render(request, 'articles/form.html', context)
 
 def detail(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
@@ -46,10 +49,17 @@ def detail(request, article_pk):
 
 @require_POST
 def delete(request, article_pk):
+    # 지금 사용자가 로그인이 되어있는가?
     if request.user.is_authenticated:
+        # 삭제할 게시글 가져오기
         article = get_object_or_404(Article, pk=article_pk)
         # if request.method == 'POST':
-        article.delete()
+        # article.delete()
+        # 로그인한 사용자와 게시글 작성자 비교
+        if request.user == article.user:
+            article.delete()
+        else:
+            return redirect('articles:detail', article.pk)
     return redirect('articles:index')
     # else:
         # return redirect('articles:detail', article.pk)
@@ -57,39 +67,44 @@ def delete(request, article_pk):
 
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            article = form.save()
-            # article.title = form.cleaned_data.get('title')
-            # article.content = form.cleaned_data.get('content')
-            # article.save()
-            return redirect('articles:detail', article.pk)
+    if request.user == article.user:
+        if request.method == 'POST':
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                article = form.save()
+                # article.title = form.cleaned_data.get('title')
+                # article.content = form.cleaned_data.get('content')
+                # article.save()
+                return redirect('articles:detail', article.pk)
+        else:
+            form = ArticleForm(instance=article)
+            # form = ArticleForm(initial={
+            #     'title': article.title,
+            #     'content': article.content
+            # })
     else:
-        form = ArticleForm(instance=article)
-        # form = ArticleForm(initial={
-        #     'title': article.title,
-        #     'content': article.content
-        # })
+        return redirect('articles:index')
     # 2가지 form 형식
     # 1. GET요청이면 초기값을 폼에 넣어서 사용자에게 던져줌
     # 2. POST요청이면 is_valid가 False가 리턴됐을 때, 오류 메시지 포함해서 사용자에게 던져줌
-    context = {'form': form}
-    return render(request, 'articles/create.html', context)
+    context = {'form': form, 'article':article}
+    return render(request, 'articles/form.html', context)
 
 @require_POST # POST요청만 있는 경우 사용가능
 def comments_create(request, article_pk):
-    article = get_object_or_404(Article, pk=article_pk)
+    # article = get_object_or_404(Article, pk=article_pk)
     if request.user.is_authenticated:
         comment_form = CommentForm(request.POST)
         if comment_form.is_valid():
                 # save 메서드 -> 선택 인자 : (기본값) commit=True
                 # commit=False : DB에 바로 저장되는 것을 막아준다
             comment = comment_form.save(commit=False)
-            comment.article = article
+            # comment.article = article
+            comment.user = request.user
+            comment.article_id = article_pk
             comment.save()
-            return redirect('articles:detail', article.pk)
-    return redirect('articles:detail', article.pk)
+            # return redirect('articles:detail', article.pk)
+    return redirect('articles:detail', article_pk)
 
 # def comments_create(request, article_pk):
 #     article = get_object_or_404(Article, pk=article_pk)
@@ -106,9 +121,18 @@ def comments_create(request, article_pk):
 
 @require_POST
 def comments_delete(request, article_pk, comment_pk):
-    comment = get_object_or_404(Comment, pk=comment_pk)
-    comment.delete()
+    # comment = get_object_or_404(Comment, pk=comment_pk)
+    # comment.delete()
+    # return redirect('articles:detail', article_pk)
+    # 1. 로그인 여부 확인
+    if request.user.is_authenticated:
+        comment = get_object_or_404(Comment, pk=comment_pk)
+        comment.delete()
+        # 2. 로그인한 사용자와 댓글 작성자가 같을 경우
+        if request.user == comment.user:
+            comment.delete()
     return redirect('articles:detail', article_pk)
+
 
 # def comments_delete(request, article_pk, comment_pk):
 #     article = get_object_or_404(Article, pk=article_pk)
